+++ test/test_helper.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
 
 require 'minitest/autorun'
+require 'open3'
 require 'json_mask'